update SStable add huffman encoding

+++ SSTable.py
@@ -3,7 +3,8 @@ from pybloom import BloomFilter
 import pickle
 import json
 from sets import Set
-
+import huffman
+import os
 
 class SSTable:
     def __init__(self, name):
@@ -56,7 +57,7 @@ class SSTable:
         self.__meta["minor"] += 1
         # Update key-file lookup table
         for key in self.__key_file:
-            if key in mem.keys():
+            if key in mem:
                 self.__key_file[key] = self.__name + table_id + "_minor"
         # Combine all minor compactions into one major compaction
         if self.__meta["minor"] == self.__meta["max_minor"]:
@@ -65,7 +66,7 @@ class SSTable:
             # Collect minor compactions and combine them to a dict
             for i in range(self.__meta["minor"]):
                 partition = self.__load_table_data(self.__name + str(i) + "_minor")
-                for key in partition.keys():
+                for key in partition:
                     mem[key] = partition[key]
             self.__dump_table(self.__name + str(table_id) + "_major")
             self.__meta["major"] += 1
@@ -89,14 +90,14 @@ class SSTable:
 
     def __dump_table(self, file_name, mem):
         ft = BloomFilter(capacity=1000)
-        for key in mem.keys():
+        for key in mem:
             ft.add(key)
         with open(file_name + "_sstable_bloom.dat", "wb") as openfile:
             pickle.dump(ft, openfile, pickle.HIGHEST_PROTOCOL)
         with open(file_name + "_sstable_keyset.dat", "wb") as openfile:
             pickle.dump(Set(mem.keys()), openfile, pickle.HIGHEST_PROTOCOL)
         with open(file_name + "_sstable_data.dat", "wb") as openfile:
-            pickle.dump(json.dumps(mem), openfile, pickle.HIGHEST_PROTOCOL)
+            pickle.dump(mem, openfile, pickle.HIGHEST_PROTOCOL)
 
     def __load_index(self, file_name):
         s = Set()
@@ -105,10 +106,18 @@ class SSTable:
         return s
 
     def __load_table_data(self, file_name):
-        data = ""
-        with open(file_name + "_sstable_data.dat", "wb") as openfile:
-            pickle.load(data, openfile, pickle.HIGHEST_PROTOCOL)
-        data = json.loads(data)
+        data = {}
+        if os.path.isfile(file_name + "_sstable_comp_dict.dat"):
+            comp_dict = {}
+            with open(file_name + "_sstable_comp_dict.dat", "wb") as openfile:
+                pickle.load(comp_dict, openfile, pickle.HIGHEST_PROTOCOL)
+            with open(file_name + "_sstable_data.dat", "wb") as openfile:
+                pickle.load(data, openfile, pickle.HIGHEST_PROTOCOL)
+            for key in data:
+                data[key] = comp_dict[data[key]]
+        else:
+            with open(file_name + "_sstable_data.dat", "wb") as openfile:
+                pickle.load(data, openfile, pickle.HIGHEST_PROTOCOL)
         return data
 
     def __load_filter(self, file_name):
@@ -116,3 +125,21 @@ class SSTable:
         with open(file_name + "_sstable_bloom.dat", "wb") as openfile:
             pickle.load(bf, openfile, pickle.HIGHEST_PROTOCOL)
         return bf
+
+    def __huffman_compression(self, file_name):
+        if os.path.isfile(file_name + "_sstable_comp_dict.dat"):
+            return
+        data = self.__load_table_data(file_name)
+        comp_dict = {}
+        for key in data:
+            if data[key] not in comp_dict:
+                comp_dict[data[key]] = 0
+            comp_dict[data[key]] += 1
+        comp_dict = huffman.codebook([(k, comp_dict[k] for k in comp_dict)])
+
+        for key in data:
+            data[key] = comp_dict[data[key]]
+        with open(file_name + "_sstable_comp_dict.dat", "wb") as openfile:
+            pickle.dump(comp_dict, openfile, pickle.HIGHEST_PROTOCOL)
+        with open(file_name + "_sstable_data.dat", "wb") as openfile:
+            pickle.dump(data, openfile, pickle.HIGHEST_PROTOCOL)
